requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
pydantic==2.11.3
pydantic-settings==2.8.1
sentence-transformers==4.1.0
faiss-cpu==1.11.0
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.6.1
pymupdf==1.25.5
python-docx==1.1.2
nltk==3.9.1
textstat==0.7.5
jinja2==3.1.6
streamlit==1.44.1
requests==2.32.3
matplotlib==3.10.1
pytest==8.3.5
httpx==0.28.1
